+++ packages/qb/pqb/src/queryMethods/select.ts
@@ -165,9 +165,12 @@ export const addParserForSelectItem = <T extends Query>(
         addParserToQuery(q.query, key, (item) => {
           const t = rel.query.returnType || 'all';
           subQueryResult.rows =
-            t === 'all' || t === 'rows' || t === 'pluck'
-              ? (item as unknown[])
-              : [item];
+            t === 'value' || t === 'valueOrThrow'
+              ? [[item]]
+              : t === 'one' || t === 'oneOrThrow'
+              ? [item]
+              : (item as unknown[]);
+
           return parseResult(rel, t, subQueryResult, true);
         });
       }