+++ packages/orm/CHANGELOG.md
@@ -1,5 +1,12 @@
 # orchid-orm
 
+## 1.6.24
+
+### Patch Changes
+
+- Updated dependencies
+  - pqb@0.10.20
+
 ## 1.6.23
 
 ### Patch Changes